feat(workflows): add build and test workflows to automate build and testing processes
The build workflow is triggered on push events to the main, env/prod, and env/dev branches, as well as pull request events of types opened and synchronize. It runs on ubuntu-latest and consists of four steps: setting up Node.js, checking out the repository, installing dependencies, and running the NestJS build.

The test workflow is also triggered on push events to the main, env/prod, and env/dev branches, as well as pull request events of types opened and synchronize. It runs on ubuntu-latest and consists of four steps: setting up Node.js, checking out the repository, installing dependencies, and running the Vitest.